moved test data to test data scripts
