+++ projects/_01_ingest/ascwds/tests/fargate/utils/test_clean_workplace_utils.py
@@ -1,6 +1,4 @@
-from dataclasses import dataclass
 from datetime import date
-from typing import Any
 
 import polars as pl
 import polars.testing as pl_testing
@@ -249,126 +247,12 @@ class TestBoundingExpressions:
         pl_testing.assert_frame_equal(returned_lf, expected_lf)
 
 
-@dataclass
-class MergeLegacyJobRoleColumnsTestCase:
-    id: str
-    mapping: dict[str, list[str]]
-    input_data: dict[str, Any]
-    expected_data: dict[str, Any]
-
-
 class TestMergeLegacyJobRoleColumns:
-    merge_legacy_job_role_columns_test_cases = [
-        MergeLegacyJobRoleColumnsTestCase(
-            id="merges_one_role_to_one_role",
-            mapping={"01": ["02"]},
-            input_data={
-                AWPClean.job_role_01_employees: 1,
-                AWPClean.job_role_02_employees: 2,
-            },
-            expected_data={
-                AWPClean.job_role_01_employees: 3,
-            },
-        ),
-        MergeLegacyJobRoleColumnsTestCase(
-            id="merges_many_roles_to_one_role",
-            mapping={"01": ["02", "03"]},
-            input_data={
-                AWPClean.job_role_01_employees: 1,
-                AWPClean.job_role_02_employees: 2,
-                AWPClean.job_role_03_employees: 3,
-            },
-            expected_data={
-                AWPClean.job_role_01_employees: 6,
-            },
-        ),
-        MergeLegacyJobRoleColumnsTestCase(
-            id="ignores_null_values_in_sum",
-            mapping={"01": ["02"]},
-            input_data={
-                AWPClean.job_role_01_employees: 1,
-                AWPClean.job_role_02_employees: None,
-            },
-            expected_data={
-                AWPClean.job_role_01_employees: 1,
-            },
-        ),
-        MergeLegacyJobRoleColumnsTestCase(
-            id="returns_null_when_all_roles_are_null",
-            mapping={"01": ["02"]},
-            input_data={
-                AWPClean.job_role_01_employees: None,
-                AWPClean.job_role_02_employees: None,
-            },
-            expected_data={
-                AWPClean.job_role_01_employees: None,
-            },
-        ),
-        MergeLegacyJobRoleColumnsTestCase(
-            id="ignores_roles_in_mapping_but_not_in_input_data",
-            mapping={"01": ["02"]},
-            input_data={
-                AWPClean.job_role_01_employees: 1,
-            },
-            expected_data={
-                AWPClean.job_role_01_employees: 1,
-            },
-        ),
-        MergeLegacyJobRoleColumnsTestCase(
-            id="creates_target_role_when_missing",
-            mapping={"99": ["01", "02"]},
-            input_data={
-                AWPClean.job_role_01_employees: 1,
-                AWPClean.job_role_02_employees: 2,
-            },
-            expected_data={
-                "jr99emp": 3,
-            },
-        ),
-        MergeLegacyJobRoleColumnsTestCase(
-            id="merges_all_matching_suffixes",
-            mapping={"01": ["02"]},
-            input_data={
-                AWPClean.job_role_01_employees: 1,
-                AWPClean.job_role_02_employees: 2,
-                AWPClean.job_role_01_starters: 3,
-                AWPClean.job_role_02_starters: 4,
-            },
-            expected_data={
-                AWPClean.job_role_01_employees: 3,
-                AWPClean.job_role_01_starters: 7,
-            },
-        ),
-        MergeLegacyJobRoleColumnsTestCase(
-            id="handles_extra_columns_in_input_data",
-            mapping={"01": ["02"]},
-            input_data={
-                AWPClean.job_role_01_employees: 1,
-                AWPClean.job_role_02_employees: 2,
-                "not_a_job_role_column": "A",
-            },
-            expected_data={
-                AWPClean.job_role_01_employees: 3,
-                "not_a_job_role_column": "A",
-            },
-        ),
-        MergeLegacyJobRoleColumnsTestCase(
-            id="handles_job_roles_with_same_characters",
-            mapping={"101": ["10"]},
-            input_data={
-                AWPClean.job_role_10_employees: 10,
-            },
-            expected_data={
-                "jr101emp": 10,
-            },
-        ),
-    ]
-
     @pytest.mark.parametrize(
         "case",
         [
             pytest.param(case, id=case.id)
-            for case in merge_legacy_job_role_columns_test_cases
+            for case in Data.merge_legacy_job_role_columns_test_cases
         ],
     )
     def test_merge_legacy_job_role_columns(self, case):
