+++ projects/_01_ingest/unittest_data/polars_ingest_test_file_data.py
@@ -1603,6 +1603,14 @@ class ApplyDataCorrectionsTestCase:
 
 
 @dataclass
+class MergeLegacyJobRoleColumnsTestCase:
+    id: str
+    mapping: dict[str, list[str]]
+    input_data: dict[str, Any]
+    expected_data: dict[str, Any]
+
+
+@dataclass
 class TestCleanAscwdsWorkplaceUtilsData:
     purge_date_test_cases = [
         CleanWorkplaceUtilsTestCase(
@@ -1797,5 +1805,111 @@ class TestCleanAscwdsWorkplaceUtilsData:
             id="empty_dataframe_are_unchanged",
             test_data=[],
             expected_data=[], # fmt: skip
+        ),
+    ]
+
+    merge_legacy_job_role_columns_test_cases = [
+        MergeLegacyJobRoleColumnsTestCase(
+            id="merges_one_role_to_one_role",
+            mapping={"01": ["02"]},
+            input_data={
+                AWPClean.job_role_01_employees: 1,
+                AWPClean.job_role_02_employees: 2,
+            },
+            expected_data={
+                AWPClean.job_role_01_employees: 3,
+            },
+        ),
+        MergeLegacyJobRoleColumnsTestCase(
+            id="merges_many_roles_to_one_role",
+            mapping={"01": ["02", "03"]},
+            input_data={
+                AWPClean.job_role_01_employees: 1,
+                AWPClean.job_role_02_employees: 2,
+                AWPClean.job_role_03_employees: 3,
+            },
+            expected_data={
+                AWPClean.job_role_01_employees: 6,
+            },
+        ),
+        MergeLegacyJobRoleColumnsTestCase(
+            id="ignores_null_values_in_sum",
+            mapping={"01": ["02"]},
+            input_data={
+                AWPClean.job_role_01_employees: 1,
+                AWPClean.job_role_02_employees: None,
+            },
+            expected_data={
+                AWPClean.job_role_01_employees: 1,
+            },
+        ),
+        MergeLegacyJobRoleColumnsTestCase(
+            id="returns_null_when_all_roles_are_null",
+            mapping={"01": ["02"]},
+            input_data={
+                AWPClean.job_role_01_employees: None,
+                AWPClean.job_role_02_employees: None,
+            },
+            expected_data={
+                AWPClean.job_role_01_employees: None,
+            },
+        ),
+        MergeLegacyJobRoleColumnsTestCase(
+            id="ignores_roles_in_mapping_but_not_in_input_data",
+            mapping={"01": ["02"]},
+            input_data={
+                AWPClean.job_role_01_employees: 1,
+            },
+            expected_data={
+                AWPClean.job_role_01_employees: 1,
+            },
+        ),
+        MergeLegacyJobRoleColumnsTestCase(
+            id="creates_target_role_when_missing",
+            mapping={"99": ["01", "02"]},
+            input_data={
+                AWPClean.job_role_01_employees: 1,
+                AWPClean.job_role_02_employees: 2,
+            },
+            expected_data={
+                "jr99emp": 3,
+            },
+        ),
+        MergeLegacyJobRoleColumnsTestCase(
+            id="merges_all_matching_suffixes",
+            mapping={"01": ["02"]},
+            input_data={
+                AWPClean.job_role_01_employees: 1,
+                AWPClean.job_role_02_employees: 2,
+                AWPClean.job_role_01_starters: 3,
+                AWPClean.job_role_02_starters: 4,
+            },
+            expected_data={
+                AWPClean.job_role_01_employees: 3,
+                AWPClean.job_role_01_starters: 7,
+            },
+        ),
+        MergeLegacyJobRoleColumnsTestCase(
+            id="handles_extra_columns_in_input_data",
+            mapping={"01": ["02"]},
+            input_data={
+                AWPClean.job_role_01_employees: 1,
+                AWPClean.job_role_02_employees: 2,
+                "not_a_job_role_column": "A",
+            },
+            expected_data={
+                AWPClean.job_role_01_employees: 3,
+                "not_a_job_role_column": "A",
+            },
+        ),
+        MergeLegacyJobRoleColumnsTestCase(
+            id="handles_job_roles_with_same_characters",
+            mapping={"101": ["10"]},
+            input_data={
+                AWPClean.job_role_10_employees: 10,
+            },
+            expected_data={
+                "jr101emp": 10,
+            },
         ),
     ]
