+++ projects/_07_workforce_characteristics/_01_starters_leavers_vacancies/tests/fargate/utils/test_prepare_utils.py
@@ -1,138 +1,21 @@
-from dataclasses import dataclass
-from typing import Any
-
 import polars as pl
 import polars.testing as pl_testing
 import pytest
 
 import projects._07_workforce_characteristics._01_starters_leavers_vacancies.fargate.utils.prepare_utils as job
-from utils.column_names.cleaned_data_files.ascwds_workplace_cleaned import (
-    AscwdsWorkplaceCleanedColumns as AWPClean,
+from projects._07_workforce_characteristics.unittest_data.polars_slv_test_data import (
+    TestPrepareUtilsData as Data,
 )
 
 PATCH_PATH = "projects._07_workforce_characteristics._01_starters_leavers_vacancies.fargate.utils.prepare_utils"
 
 
-@dataclass
-class ReduceToPublishedRolesTestCase:
-    id: str
-    mapping: dict[str, list[str]]
-    input_data: dict[str, Any]
-    expected_data: dict[str, Any]
-
-
 class TestReduceToPublishedRoles:
-    reduce_to_published_roles_test_cases = [
-        ReduceToPublishedRolesTestCase(
-            id="merges_one_role_to_one_role",
-            mapping={"01": ["02"]},
-            input_data={
-                AWPClean.job_role_01_employees: 1,
-                AWPClean.job_role_02_employees: 2,
-            },
-            expected_data={
-                AWPClean.job_role_01_employees: 3,
-            },
-        ),
-        ReduceToPublishedRolesTestCase(
-            id="merges_many_roles_to_one_role",
-            mapping={"01": ["02", "03"]},
-            input_data={
-                AWPClean.job_role_01_employees: 1,
-                AWPClean.job_role_02_employees: 2,
-                AWPClean.job_role_03_employees: 3,
-            },
-            expected_data={
-                AWPClean.job_role_01_employees: 6,
-            },
-        ),
-        ReduceToPublishedRolesTestCase(
-            id="ignores_null_values_in_sum",
-            mapping={"01": ["02"]},
-            input_data={
-                AWPClean.job_role_01_employees: 1,
-                AWPClean.job_role_02_employees: None,
-            },
-            expected_data={
-                AWPClean.job_role_01_employees: 1,
-            },
-        ),
-        ReduceToPublishedRolesTestCase(
-            id="returns_null_when_all_roles_are_null",
-            mapping={"01": ["02"]},
-            input_data={
-                AWPClean.job_role_01_employees: None,
-                AWPClean.job_role_02_employees: None,
-            },
-            expected_data={
-                AWPClean.job_role_01_employees: None,
-            },
-        ),
-        ReduceToPublishedRolesTestCase(
-            id="ignores_roles_in_mapping_but_not_in_input_data",
-            mapping={"01": ["02"]},
-            input_data={
-                AWPClean.job_role_01_employees: 1,
-            },
-            expected_data={
-                AWPClean.job_role_01_employees: 1,
-            },
-        ),
-        ReduceToPublishedRolesTestCase(
-            id="creates_target_role_when_missing",
-            mapping={"99": ["01", "02"]},
-            input_data={
-                AWPClean.job_role_01_employees: 1,
-                AWPClean.job_role_02_employees: 2,
-            },
-            expected_data={
-                "jr99emp": 3,
-            },
-        ),
-        ReduceToPublishedRolesTestCase(
-            id="merges_all_matching_suffixes",
-            mapping={"01": ["02"]},
-            input_data={
-                AWPClean.job_role_01_employees: 1,
-                AWPClean.job_role_02_employees: 2,
-                AWPClean.job_role_01_starters: 3,
-                AWPClean.job_role_02_starters: 4,
-            },
-            expected_data={
-                AWPClean.job_role_01_employees: 3,
-                AWPClean.job_role_01_starters: 7,
-            },
-        ),
-        ReduceToPublishedRolesTestCase(
-            id="handles_extra_columns_in_input_data",
-            mapping={"01": ["02"]},
-            input_data={
-                AWPClean.job_role_01_employees: 1,
-                AWPClean.job_role_02_employees: 2,
-                "not_a_job_role_column": "A",
-            },
-            expected_data={
-                AWPClean.job_role_01_employees: 3,
-                "not_a_job_role_column": "A",
-            },
-        ),
-        ReduceToPublishedRolesTestCase(
-            id="handles_job_roles_with_same_characters",
-            mapping={"101": ["10"]},
-            input_data={
-                AWPClean.job_role_10_employees: 10,
-            },
-            expected_data={
-                "jr101emp": 10,
-            },
-        ),
-    ]
-
     @pytest.mark.parametrize(
         "case",
         [
             pytest.param(case, id=case.id)
-            for case in reduce_to_published_roles_test_cases
+            for case in Data.reduce_to_published_roles_test_cases
         ],
     )
     def test_reduce_to_published_roles(self, case):
